ml_experiments/invoice_classifier/requirements.txt
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
mlflow==2.13.0
imbalanced-learn==0.12.2
shap==0.45.1
fairlearn==0.10.0
pyyaml==6.0.1
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
pytest==8.2.0
